requirements.txt
# Core dependencies
langchain==0.1.0
langchain-openai==0.0.2
openai==1.12.0
tiktoken==0.5.2

# Vector database
chromadb==0.4.22

# PDF processing
pypdf2==3.0.1
fpdf==1.7.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
typing-extensions>=4.5.0
numpy>=1.24.2
pandas>=1.5.3

# Logging and monitoring
logging>=0.5.1.2

# Optional but recommended
pytest>=7.3.1  # For testing
black>=23.3.0  # For code formatting
isort>=5.12.0  # For import sorting
mypy>=1.3.0   # For type checking

# For Chinese language support
fonttools>=4.39.3